[OEUI-226] Add a way to see repeated documents (#653)
## Description

Closes OEUI-226

This PR adds a button to any CorpusLInk document that has repeated
documents. Clicking the button will take over the CorpusLink drawer or
tab and show the repeated documents.

This is a draft because I'd like to figure out if we want the scroll
position to be maintained when going back to the main CorpusLink content
and I need to add tests. I also need to figure out how to reset the
selected repeated document when navigating to another thread. May just
need to add on to what we have in the current thread loader.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally, will add some kind of test

## Output / Screenshots


https://github.com/user-attachments/assets/aeeb1656-34ce-4bb0-94b5-b8daf7ac64cb



https://github.com/user-attachments/assets/fd81c1ed-84be-4470-b13c-d47631b91b43


### A11y

I've made sure interactive elements have proper labels and elements