[OEUI-320] Fix Navigating to a document from the dataset explorer shows an error (#675)
## Description

Notice when navigating to a specific document it shows an error. This PR
fixes it.

## Reviewer Instructions

Notice that the error is cause if the total from meta property is > 1
then it throws error. Previously we were using es but now we switch to
infini_gram so this can be the case.

## Testing Plan

Verify UI works. If CI passes we are good.

## Output / Screenshots


https://github.com/user-attachments/assets/9e83de4d-2967-4ab9-b290-8052cf9e24a5


### A11y

N/A