[OEUI-325] Fix abort button not working (#686)
## Description

The abort button not working this PR fixes it. In our [previous
PR](https://github.com/allenai/olmo-ui/pull/592) we forgot to pass the
event handler to the component so it causes the component to not know
anything about it.

## Testing Plan

Verify UI works. If CI passes we are good.

## Output / Screenshots


https://github.com/user-attachments/assets/e735eb5a-cb44-49f1-8917-5f063ba1a841


### A11y

N/A