[Unplanned] Combine fake app contexts for testing (#682)
## Description

I had a little time after reviewing @yensung 's PR and wanted to see if
this was a viable approach. Turns out it is!

This PR replaces the second fake app context we made with some
modifications to the first one. The changes allow us to pass in actions
with getters and setters, letting us mock actions in the store that
actually affect things.

The format to make a mock action is a _little_ funky since it needs the
`set` from the state creator closure, but it works!

## Testing Plan

Tested by removing the streaming message ID set call and making sure
tests still pass.

---------

Co-authored-by: yensung <george129.chen@gmail.com>