[OEUI-278] [OEUI-286] Pin scroll to the bottom after user scrolls to see thread contents (#672)
## Description

Closes OEUI-278, OEUI-286

This PR changes the behavior of our thread display so that if a user has
scrolled to the bottom, the scroll will stick to the bottom while a
message is streaming in.

Note that I've changed the streaming indicator in this PR as well.
Before it was showing up below the message, which didn't fit the design!
It was also messing with the auto-scrolling.

## Testing Plan

Added screenshot tests with Playwright.

## Output / Screenshots


https://github.com/user-attachments/assets/839171a9-7006-4c25-a62f-33cf4db06f89

### A11y

I've ensured we only smooth scroll if the user has noted no preference
for reduced motion. Buttons are properly labeled.