[OEUI-374] make T&C only show up after someone's logged in (#703)
## Description

Closes OEUI-374

I included some changes to NewApp so we don't have the useEffect
handling loading, moving the data loading to react-router's loaders as
we have been in the past.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots



https://github.com/user-attachments/assets/6219ac8e-e60e-43a7-96d1-d613ae0d2b22



https://github.com/user-attachments/assets/32ace3d7-fb6f-479d-9277-9afd82ba91b2