[OEUI-375] Add legal notice below prompt for users that have not accepted the T&C (#706)
## Description

This adds the term language for non auth'd users who haven't agreed to
the T&C.
This one doesn't really do much on its own, and goes along with the T&C
dialog only showing when logging in. PR: #703

The UI is likely not final, but just a PoC

This also adds the accepting value into zustand while saving it remotely
(following the pattern getUserInfo used) so that they agreement gets
set, and the slice gets gets updated together. If this need to be done
in two steps, let me know.

## Reviewer Instructions

This needs #703 to be of any real use, as the current main requires
acceptance of the T&C initially.

## Testing Plan

Manually verified

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/38c40aed-b7c4-4ac4-83b0-76ce68442539)

### After

Desktop:

![image](https://github.com/user-attachments/assets/b9593a3c-5fd0-42a8-bd3e-eca467fd7f10)

Mobile:

![image](https://github.com/user-attachments/assets/03636756-e684-4405-a26b-6e78e068ff43)


### A11y

Should be fine with semantic