[OEUI-390] keep the input enabled but prevent submitting prompts (#707)
## Description

Closes OEUI-390

This is a change from what we previously had. I'm keeping the prompt
enabled but preventing submission while we're streaming. This makes the
focus handling much easier! Since we're not disabling it, we don't need
to refocus it after the stream is done. Since we're not moving focus
around on our own, we prevent any UX issues from someone having focus on
something else and getting interrupted.

## Testing Plan

Local, unit tests

## Output / Screenshots
(ignore the error at the end, that's from our mocks)


https://github.com/user-attachments/assets/f2e06048-3412-4083-b2fe-7bd5db13efdf

### A11y

This PR tries to keep things keyboard-user friendly by not stealing
their focus.