Refactor captcha handling to not block first submission (#716)
This PR fixes a bug so that we don't need to submit twice to get a
submission. Previously we would have to submit to get a token, not see
anything in the UI, then resubmit to get an answer.

There may be a lighter touch way to handle this but I couldn't figure it
out!