Adding proper markdown component overrides for elements (#732)
## Description

Adding custom `<hr>` and moved the link and paragraph customizations to
custom components as well.

The changes to `<p>` and `<a>` should not be visually different, just
styles moved to component. `<hr>` treatment is updated.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

Link:

![image](https://github.com/user-attachments/assets/f1c8fd9f-0666-4724-ba62-9bcc28992dcb)

`<hr>` is `black` or `white` with `0.25` opacity and has `1.5rem`
top/bottom margin:


![image](https://github.com/user-attachments/assets/08067398-e491-41cf-bf26-7a834023c572)

![image](https://github.com/user-attachments/assets/77cdbc91-f29e-4142-b087-d764acaf0a9c)


### A11y

How do keyboard navigation and screen readers work with this PR?