[OEUI-386] Auto-size the prompt input (#726)
## Description

Closes OEUI-386

Used this solution to get the multi-line prompt working:
https://css-tricks.com/the-cleanest-trick-for-autogrowing-textareas/

## Reviewer Instructions

Is this understandable? let me know if i should add any comments
anywhere. how do the design choices look?

## Testing Plan

Tested locally, make sure everything passes with existing tests.

## Output / Screenshots

Unfocused:

![image](https://github.com/user-attachments/assets/079f88c2-821b-481b-90bf-736a1fbb485f)

![image](https://github.com/user-attachments/assets/d65a8353-f18e-4f95-9508-b90f867bb4c3)


Focused (turns out `:focus-visible` is always true if you're focused on
an input. @aaronsarnat we can make it so it only shows with keyboard,
but it'll be annoying):

![image](https://github.com/user-attachments/assets/0029157b-476f-401e-a65e-c4adf0eef086)

![image](https://github.com/user-attachments/assets/a99061dd-6b30-4187-bf74-7954cc0019cb)


The input getting bigger:


https://github.com/user-attachments/assets/19454f56-5456-430e-b1c3-adf612f155b5

<details>
<summary>
With error message: 
</summary>


![image](https://github.com/user-attachments/assets/d6d20867-dcac-49cd-baee-2e45ce49f854)


![image](https://github.com/user-attachments/assets/eb26b7ff-c692-4463-aadf-f12a94373df8)

</details>

<details>
<summary>
With a lot of lines:
</summary>


![image](https://github.com/user-attachments/assets/46629a06-cff5-4492-8483-579909fc22b7)

</details>

<details>
<summary>
With a lot of lines on mobile:
</summary>


![image](https://github.com/user-attachments/assets/7db40563-925f-4af5-8298-02233d9f5beb)

</details>

<details>
<summary>
One line on mobile:
</summary>


![image](https://github.com/user-attachments/assets/80fe898e-25cd-4605-996d-379701924d9e)

</details>

<details>
<summary>
Submission and stop video:
</summary>


https://github.com/user-attachments/assets/5fb28e67-c5dd-4329-b117-1400a0493172

</details>


### A11y
Labels and focus states have been set up correctly. The submit and stop
button are properly focusable.