little more padding -- smaller icons (#767)
## Description

Just makes the icons a bit smaller

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots


![image](https://github.com/user-attachments/assets/c30bdcb8-157e-4f06-8e78-a19a2b175488)

### A11y

How do keyboard navigation and screen readers work with this PR?