[OEUI-419] Allow users to pre-select a model through the query param (#774)
## Description

Closes OEUI-419

This PR makes it so that people can share links with a pre-selected
model. This can help eliminate the need for churn in our config to set
the order for new releases.

It uses a `model=<model-id>` search param to handle this. for example,
`https://playground.allenai.org/?model=OLMo-peteish-dpo-preview` would
load the peteish model.

Notably, this query param checking only applies to the root Playground
page. Selected threads will still have the model used for the last
request selected.

## Testing Plan

Tested locally, added a unit test

## Output / Screenshots


![image](https://github.com/user-attachments/assets/1462b73e-6828-4e1b-b9ab-949a3e092a90)


![image](https://github.com/user-attachments/assets/68e9e981-f937-456e-8b9a-453d2a54a801)


![image](https://github.com/user-attachments/assets/83f1bc71-191f-4473-9a96-cbd8b1713af6)