[OEUI-453] Allow model selection queries to select deprecated models (#779)
## Description

Closes OEUI-453

This PR makes it so that users can select a deprecated (or just hidden)
model through the query param. It also makes it so that we'll set the
selected model to the deprecated one if we load a thread that used a
deprecated model.

I also did some refactoring of the ModelSelectionDisplay, including
renaming it to ModelSelect.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally, added e2e and unit tests

## Output / Screenshots


https://github.com/user-attachments/assets/79181da7-1518-44b0-84af-79159f7483d0