Fix the alert again (#793)
We changed the legal text so now the alert's broken! I fix it here by
checking for the icons that link to the homepage.

We also had bad alt text on our SVG icons. This fixes that and enforces
passing a title to icons through TS to give us some more checking around
that.

---------

Co-authored-by: Huy Tran <84343285+huytr1995@users.noreply.github.com>