Support multiple indexes in Dataset Explorer (#817)
## Description
Closes https://github.com/allenai/playground-issues-repo/issues/100

This PR makes the Dataset Explorer and Dolma API support getting
documents for any index, not just one. It's backwards compatible through
defaulting to an index if the index isn't set in the path.

This also changes the CorpusLink cards to link to the right index when
opening a document in the dataset explorer.