log in again if the refresh token is missing (#823)
## Description

This is some functionality we missed when switching to allow unauth
users. If a user is missing their refresh token we just ignore it and
treat them as an anonymous user right now. This change will make it so
we re-authenticate them and get the new refresh token. This should just
flash a few screens as a redirect and not require a new login since
they're already authenticated.

## Testing Plan

Tested locally.