Add file upload button (#819)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/4

This PR adds a button that handles file uploads, hidden behind a feature
flag.

Not terribly concerned about how it looks since we'll probably go
through another design thing in the future. Just tried to make it look
decent.

It's currently restricted to just one image. We may want to make that
configurable per model in the future.

It also always shows if the feature flag is enabled. We may want to
either hide or disable it in the future based on the model we're using.
That would require some API changes, either to add a new model type or
some new fields.

## Testing Plan

Tested locally.

## Output / Screenshots

![image](https://github.com/user-attachments/assets/4b542b37-4d65-4fb2-af71-f7e289000ab5)

focused: 

![image](https://github.com/user-attachments/assets/11391fa6-561e-48ca-bbd6-c4a067127943)



https://github.com/user-attachments/assets/6aaec352-c57d-499b-8294-ade576f98cc9


### A11y

I've ensured that appropriate roles are set for the label, focus styling
is present when needed, labels are set correctly, and the hidden input
isn't focusable.


![image](https://github.com/user-attachments/assets/f7b20927-00f3-4635-badf-e82b778dbae1)