fix issues coming from re-authenticating when getting a new access token (#831)
## Description

I broke some auth flow stuff with the last PR to log people back in if
their refresh token was bad. This PR attempts to fix those issues.

The main issue that was happening (from what I could tell) was that we
were getting user auth info from the root layout on all of our pages.
This would cause an infinite loop of logging in and re-authing since we
were triggering a re-auth on the `login-result` page, which sets the
data we need.

I fixed this issue by moving the login pages outside of the routes that
request user info. This seems to have fixed the issue on my local.

## Testing Plan

Tested locally by invalidating and deleting my refresh token then
refreshing the page. I was able to reproduce errors before by deleting
my refresh token and refreshing the page.