Stable scroll gutters on thread page (#826)
## Description

This finally bugged me enough to do something about it.

When we load a thread, we get weird-looking thread gutters and layout
shift because `ThreadDisplay` has `overflow: scroll` and
`ThreadPlaceholder` didn't.

I've fixed this by forcing a horizontal scroll in the container and
making the scroll bars look nicer.

Note that `ThreadPlaceholder` and `ThreadDisplay` have the same code
that isn't actually shared. I've made a ticket to have them share a
layout in the future. I started it here but it turned into a bit more
than I wanted to bite off.

## Testing Plan

Tested locally

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/97d6acbe-61f0-4306-ba8f-fea4defcbd28)


![image](https://github.com/user-attachments/assets/1b57a977-f930-4cbe-b0ff-4aeda8640152)

### After


![image](https://github.com/user-attachments/assets/f1844fdc-689c-42fc-87c8-263bb7b3b68a)

![image](https://github.com/user-attachments/assets/9be1d237-12dc-4a49-896e-a6ae5ca04444)