Show file thumbnails and allow users to remove them (#828)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/110

This PR adds a thumbnail preview of images the user has chosen to
upload. It also allows users to remove those images by clicking a
button.

To style the close button more easily I set up SVGR to allow us to
import svgs as react components.

## Testing Plan

Tested locally, added a unit test

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/55424af6-5b33-4293-99aa-2e261c76491e)

(the colors on this are old but i didn't want to re-record)



https://github.com/user-attachments/assets/4cab47de-91e3-4cb5-817b-bd44fe64e602



### A11y

I've ensured that the buttons are focusable, images have appropriate alt
text, and the button has an appropriate label.