Allow both span and document selection (#835)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/112

This PR makes it so users can see a span to see corresponding documents
and select a document to see corresponding spans.

To accomplish this, I modified the selection stuff in the
`AttributionSlice`. Instead of selected documents or spans at the root,
I made a new "selection" property that holds a "type" that can either be
"span" or "document". This lets us provide more info about _what_
selection the user did and more easily reset the selection state for the
other type.

I also combined the `documentFirst` and `spanFirst` selectors. I think
there may be a better solution that doesn't involve a bunch of markdown
re-rendering by modifying the `AttributionHighlights` to do the
document/span logic instead, but I thought that was going to be too much
for this already kinda big PR!

## Testing Plan

Tested locally. Tested with existing unit and e2e tests.

## Output / Screenshots




https://github.com/user-attachments/assets/a59d801a-0574-4c45-90a1-776eaff98b61



### A11y

How do keyboard navigation and screen readers work with this PR?