Add modal to explain training text matching (#841)
## Description

Closes: allenai/playground-issues-repo#114

Adds a Modal to explain the what "Training text matches" are

I don't have a design for the modal, so this is just reusing a modal we
have. Should be easy to update this.

This also doesn't address the design changes in the sidebar, as I think
there are current PRs out there that are either addressing them or
should come in first so that the whole can be styled together.

## Testing Plan

Updated one test looking for the text `CorpusLink` which was changed to
`Training text matches`. All tests pass locally.

## Output / Screenshots

### Before

no modal

### After

#### Sidebar
*Dark*

![image](https://github.com/user-attachments/assets/6dc4278f-713d-4fc0-8d83-5b9c47603a3f)

*Light*

![image](https://github.com/user-attachments/assets/a12788ac-36f3-461e-8107-52a8a21342e2)


#### Modal
*Dark*

![image](https://github.com/user-attachments/assets/bfe54474-c336-4f8c-98a2-47ec92eb7fe7)

*Light*

![image](https://github.com/user-attachments/assets/dd9fc4aa-177f-4372-9138-288070f266d1)


#### Mobile

![image](https://github.com/user-attachments/assets/9ecaffc0-043e-4d73-8d3f-06ad39f63318)


### A11y

How do keyboard navigation and screen readers work with this PR?