Move snippet fetching up one level so we can use it sooner (#842)
## Description

Closes allenai/playground-issues-repo#121

This dynamically makes the `Locate span` plural. This just moves the
logic that handles gathering the snippets up one level in the hierarchy,
as this code was being run inside `AttributionDocumentCardSnippets`
anyways.

I don't have a URL to test this against, I just used testing env. If
someone has a message ID that would be awesome to test against.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Added an `expect` to a test that should pass fail based on this.

Passing locally.

## Output / Screenshots

### Before

`"Locate span(s)"`

### After

`/Locate spans?/`
;)

### A11y

How do keyboard navigation and screen readers work with this PR?