add a button to select and deselect a message to show in the CorpusLink drawer (#839)
## Description

The functionality of this is a bit of a WIP but we didn't want to block
this. We'll be figuring out how we want the drawer to react to
showing/hiding things and dialing that in in another PR.

@aaronsarnat asked for the text button to be changed to an icon button.
I was planning on having this be an instance of a varnish-ui button but
I'm skipping that because we have this exact button component made
already!

## Testing Plan

Tested locally. Not adding tests just yet because functionality isn't
finalized.

## Output / Screenshots

Button selected state:

![image](https://github.com/user-attachments/assets/609108b0-352d-491e-a05e-d1d214fc6d0b)

Button unselected state:

![image](https://github.com/user-attachments/assets/db62d59f-5c8a-4edb-8537-0fab7a0e9492)



https://github.com/user-attachments/assets/57395c1a-8416-461f-90a6-3747bff3b3a6

### A11y

The new button has appropriate tooltips and labels applied