Stop automatically selecting followup messages (#837)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/122

We probably should've done this after the one that lets us select
messages, I'm going to block this until that's done! You can't really
use CorpusLink with this PR until it's done lol

We still automatically select the latest message in a thread that we
start loading.

## Testing Plan

Tested locally. Going to add more tests when we implement the message
selection buttons.

## Output / Screenshots


https://github.com/user-attachments/assets/2bc6cdd7-f11d-4f77-a4a6-b08cb83c2807