FIx the padding on the prompt button (#853)
## Description

Fixes the padding on the button

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/1bd186a3-f0da-41e1-be9d-82694e73da6b)

### After


![image](https://github.com/user-attachments/assets/b6b01edd-66da-46fa-a488-7ce135014051)

### A11y

How do keyboard navigation and screen readers work with this PR?