clear span/document selections when the selected message changes (#859)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/158

I intentionally didn't clear selections when the drawer is closed since
we're making the drawer only control itself.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally. Tried to make a test but it was kinda tough? We need to
figure out how to test store actions better, the zustand docs don't
help.

## Output / Screenshots


https://github.com/user-attachments/assets/117d22bc-f6e0-4533-8b00-138bcbe98267