Auto hide controls on desktop + cursor (unless last message) (#891)
## Description

Shows the controls for the last message in a thread, and for messages
before, it displays on hover over the message.

## Reviewer Instructions

Signed off by @aaronsarnat 

Related: allenai/playground-issues-repo#130

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

Apologizes for the potato video, I think I exported it after already
saving it low res, so its ultra low res.


https://github.com/user-attachments/assets/48c4f7ef-6b27-4ba5-9db0-25a1aad6b643

For reference, here is the same pattern used by Gemini and Meta



https://github.com/user-attachments/assets/a2ca9022-fc40-43ab-bbef-00d6485a1792


https://github.com/user-attachments/assets/d73efdf8-8bb3-47d0-a538-86b9830058da




### A11y

How do keyboard navigation and screen readers work with this PR?