Show the CorpusLink button as text for destkop (#892)
## Description

Changes the button to show text matches to be text (as it was
previously). It remains as a button for mobile.

This also adds hover color to the text.

allenai/playground-issues-repo#130

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots



https://github.com/user-attachments/assets/153a0c1a-e9e9-4935-8064-21f5d5ab13b9

### A11y

How do keyboard navigation and screen readers work with this PR?