Show/Hide drawer when selecting or locating a span (#893)
## Description

**On mobile:**

Opens the drawer when selecting a span. Closes the drawer when "Locating
a span"

## Reviewer Instructions

This is a basic addition, I wonder if it should scroll the select into
view when clicking `locate span`. I still see this as an iterative UX
improvement, as we are not doing anything right now, which is more
confusing than the span being out of the view, (which it would still be
even if you manually closed the drawer).

## Testing Plan



## Output / Screenshots

### Before



https://github.com/user-attachments/assets/2e5d5b57-1400-48a4-99b2-b7ebbb591d28




### A11y

Keyboard regarding this PR seems fine, but it caused me to notice we
have to keyboard navigation on spans.