Fix overlapping spans (#930)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/156
Depends on https://github.com/allenai/olmo-api/pull/259

This PR fixes the overlapping span issue. 

The issue was caused by two things:
* Our regex was global, making it replace every instance of a span
instead of just the first
* Our regex didn't match the current state of attribution highlights,
which now use HTML elements instead of Markdown directives.

I've addressed both of those issues in this PR.

## Testing Plan

Tested locally. Added mocks for a response that has this problem.

## Output / Screenshots

### Before

I forgot to take a screenshot of the before state of the test response I
have and it's hard to get now so you'll have to trust me on this one

### After


![image](https://github.com/user-attachments/assets/c7d88000-89c1-4c53-bcf9-54abe09d536a)