Fix issues with analytics when `window.heap` doesn't exist (#966)
## Description

We got a report in our Discord that a user wasn't able to select models
because tracking was failing. This makes sure our tracking setups exist
before we call them.

I've also wrapped the `track` fn in a try/catch so this can't happen
even if we _do_ break it's handled.