Implement new route for Model Config (#965)
## Description

Closes:  https://github.com/allenai/playground-issues-repo/issues/327

## Reviewer Instructions

I made comment about how we should set up some sort of flag in the user
object to make sure if we have the right permission we would allow them
to go to the route but no just throw 401 unauthorized. Also make a note
that for now redirect `/admin` to `/model-configuration` until we figure
out what we want to do with /admin.
## Testing Plan

Just make sure the route loads and redirect to model configure ui.

## Output / Screenshots


https://github.com/user-attachments/assets/581d4728-de8f-4412-bd04-dfda8612c952


### A11y

N/A