Set up openapi-react query and make an example (#970)
This PR sets up `openapi-react-query` and sets up an example of how I
think we should use it for the models page.

This includes setting up `openapi-fetch`, `openapi-typescript`,
`openapi-react-query`, and `@tanstack/react-query`. We're using those in
conjunction with `react-router` `loader`s and `action`s to handle
loading and modifying models on the `/test` page.

I don't plan on merging this PR until we have a chance to talk about it.
I'll also need to move the example to the model config page that
@huytr1995 set up so it's not available in prod.