Add ModelConfigurationPage (#972)
## Description

Closes: https://github.com/allenai/playground-issues-repo/issues/328
Figma:
https://www.figma.com/design/a5xG9ZSnNO29c7uiDQ2fLk/Playground-2025?node-id=463-400&m=dev

Add new model list component to display all available model from the
back end.

## Reviewer Instructions

@mtblanton per discussion we cant do the click button to enable the drag
and drop for now. This is because dragAndDropHooks either has to be
undefined or has value and cant be reinitialized again. I use GridList
from react-aria-components to render the list and make 2 stories book
component for it ( we will port it to varnish ui soon). The rest is
varnishui and pandacss styling.

## Testing Plan

Verify UI works. If CI passes we are good.

## Output / Screenshots


https://github.com/user-attachments/assets/8b8fc8c8-fc5a-4731-be0c-d3673735b357


### A11y

N/A