Add Panda breakpoints (#1010)
I was setting this up for the scroll to bottom. I didn't end up using it
(for now), but I thought these would be nice to have. They match the
current MUI `@media` breakpoints and `@container` sizes

I'd prefer to use container sizes, and I think the comparison might be a
good spot (as we will have the UI that is 1 column and 2 column), but I
don't know if our 2 sizes are sufficient for this. I'm thinking we
should add more.


NOTE:
`desktop` can be removed if it seems weird, I added it to correspond to
`DESKTOP_LAYOUT_BREAKPOINT` which is `lg`