Update scroll to bottom design (#1011)
This is an initial design update for the scroll to bottom.

Closes: allenai/playground-issues-repo#347

I can't say that I'm pleased with the mobile, the problem is the width
of the MessageList and Prompt diverge below the max-width 750, with
MessageList getting padding from two places and the prompt one.

This might be easier to fix when doing the thread list updates.

These should probably be container queries also, but as noted in the
other PR, I don't think we have final sizes. So this keeps it
"shippable" for now.

<img width="1095" alt="image"
src="https://github.com/user-attachments/assets/316754be-36b1-47d8-b17a-ec083b4339dc"
/>
<img width="393" alt="image"
src="https://github.com/user-attachments/assets/ac75298f-f838-40df-9c0d-9e5ab06d8eb2"
/>