Model selection, loading, split Select component (#1015)
## Description

How this all goes together is subject to change, but my initial thoughts
are to have the concept of a `threadVIewId` (currently its essentially
and `idx`), This is used to store the selected model (as we may not have
a root thread id).

The model selection for comparison is stored as UI state in a slice.

This has some compatibility code to keep the current single thread view
from breaking, but I'd expect that to get replaced (comparison with a
single thread should work the same as the current thread view)

Related: allenai/playground-issues-repo#397
Closes https://github.com/allenai/playground-issues-repo/issues/397
Closes https://github.com/allenai/playground-issues-repo/issues/398

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>