single `useThead` hook with selector (#1033)
I've updated the to use a single query function `useThread` that takes a
selector.

It will use imperative fetches when needed (page loader).

I was hoping to split out the OLMoTrace changes, but they relied on the
changed hooks, so It would be two broken PRs instead of a single working
one, the OLMoTrace stuff is mostly the same as the other PR.

Closing #1028 in favor of this.

Closes: allenai/playground-issues-repo#410