Update TheadSlice to use new type (#1051)
## Description

Changes the HistoryDrawer to use the new types from the `v4-handlers`.
This should really be changed over the react-query, but I had some
issues, and found this to be the easier route at the moment.

Everything is switched over the be `Thread` based and the code for the
drawer updated to work with those shapes instead.

---------

Co-authored-by: stevena <stevena@allenai.org>