Context-based DI + stream event handlers for adding stream features in their relevant Containers (#1068)
A lot of changes! 

The last few features were tricky, like not breaking the single-vs-multi
thread abstraction for the "scroll to new user message" feature. I
considered a few approaches and added a way to register streaming event
handlers. I think that will really help develop new stream-related
features in the actual containers that correspond to those features!
Some features are still implemented in the stream processing itself, but
they *should* be able to be move to Containers. To test out the pattern
I moved the "form reset on first message" out to QueryFormContainer.
Plenty of other features could be implemented that way.

At this point a lot of files can be carefully deleted! I don't think
anything is *really* using ModelSlice or CompareModelSlice for state
management. That is, unless my searching missed something. There were a
lot of places that needed changed. Many of them required slightly
different approaches to using the new non-global state.

A couple of the E2E tests would fail intermittently. So I increased
their timeouts. Otherwise E2E tests are unchanged. So, behavior after
the big refactor should be pretty close to the way it was. Unit tests
for Comparison are currently failing. I've been focussed on feature
parity for single-thread, but also trying to think through how it will
be handled for comparison.

---------

Co-authored-by: jonathan m borchardt <jonathanb@allenai.org>