Expandable system prompt input (#1096)
## Description


https://github.com/issues/assigned?issue=allenai%7Cplayground-issues-repo%7C444

Adjusted the system prompt and description fields on the edit modal
admin to be more usable. They can now expand to show a larger area.


### Notes:

##### Modal
Styling the modal with `color='reversed'` broke the styling of content
inside of the modal, so it is left in light mode. When varnish is
updated to the latest version, this issue should be resolved.

##### Textarea

There seems to be a bug in varnish textarea where when given a long
string of text (with no new lines) the size of the textarea grows. This
is an issue that likely needs to be fixed in varnish. Not a blocker but
is unexpected behavior.

I did not end up making the text area resizable due to some weird
behavior related to the issue above. When fixed can add the resize
vertical property.


## Reviewer Instructions
Open the modal config in the admin view, update the description in both
the textarea and modal. Confirm that the information is saved.

## Testing Plan

Open the admin modal config and change description and system prompt.

## Output / Screenshots

### Before

<img width="389" height="950" alt="Screenshot 2025-07-29 at 3 19 55 PM"
src="https://github.com/user-attachments/assets/306678ab-0196-4b2c-88d8-ce4fdc3fd684"
/>


### After
<img width="1237" height="934" alt="Screenshot 2025-07-29 at 3 17 00 PM"
src="https://github.com/user-attachments/assets/11dd2ad3-e082-4ce4-a2e4-7ec3a4e71ba8"
/>
<img width="1454" height="945" alt="Screenshot 2025-07-29 at 3 52 09 PM"
src="https://github.com/user-attachments/assets/cd8ca975-8d7e-4f24-bde4-fb6ef57b4523"
/>



### A11y

How do keyboard navigation and screen readers work with this PR?

Keyboard navigation still works as expected.

---------

Co-authored-by: Caleb Ouellette <calebo@allenai.org>