Update parameter drawer (#1129)
## Description

This Pr adds stop words and max tokens back into the drawer. 

I used the existing components and existing copy for this PR.

closes
https://github.com/issues/assigned?q=is%3Aissue+state%3Aopen+archived%3Afalse+assignee%3A%40me+sort%3Aupdated-desc+org%3Aallenai

If anyone has context on why this was removed that might be good to
know? Everything seemed to work fine when i added it back in.

While testing this i noted some weird behavior, we don't send options on
the first message. I'm going to look into fixing this before merging.

This Pr also sets the max tokens to 2048, is that a good value?

settings are not sent unless the user open the side drawer once. This
probably doesn't matter if our defaults are the same on frontend and
back, but it is a little odd.

## Reviewer Instructions

Play around with options in drawer ensure they are send to backend on
api call.

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before


<img width="721" height="796" alt="Screenshot 2025-08-14 at 4 33 28 PM"
src="https://github.com/user-attachments/assets/b5631cf6-b6a1-4d6e-bb27-316d69348085"
/>

### After

<img width="695" height="706" alt="image"
src="https://github.com/user-attachments/assets/ade7387b-eb43-4641-9095-edf4fde777c2"
/>


### A11y

How do keyboard navigation and screen readers work with this PR?