Collapsible varnish component (#1111)
## Description
Closes https://github.com/allenai/playground-issues-repo/issues/521

This is the varnishui component for collapse. Individual wigets _should_
be able to be easily made from here.

## Reviewer Instructions

I made this in storybook to make it easy, but I expect to move it to
varnish-ui (thats why its in a varnish subdir).

When moved and package updated, this directory can go away.

I included a non composed version, as thats what we have typically done
with varnish-ui -- but I'd expect the composable version is what we are
really interested in.

This only includes a dark version of this component, as thats the only
designs I hade (I've request light designs from Karen).

React Aria Components uses `hidden="until-found"` which makes animating
an issue. I've included chrome only styles while we evaluate wether RAC
will work out of the box for this, or if RAC useHook needs to be used
(or something else).

## Output / Screenshots


https://github.com/user-attachments/assets/9220f3af-d66b-4df4-bae6-ffdeeeb7ca61

### A11y

using RAC default behavior