Handle new streaming chunk types (#1125)
## Description
Closes https://github.com/allenai/playground-issues-repo/issues/516

This PR makes it so our stream handling knows how to handle the new
stream chunk types. This means that we're in a good spot to start
handling the UI for thinking and tool calling!

I did make some changes in the API to make these schemas work. ~I'll
make a PR for that.~ Here's a PR for those changes:
https://github.com/allenai/olmo-api/pull/358

## Reviewer Instructions

I've used `immer` to handle the streaming updates since they're deeply
nested. What do you think of that?

## Testing Plan

Tested against the Test Model against @CalebOuellette 's API PR. It's
handling thinking and tool call chunks appropriately as well as the
final result from the API.

---------

Co-authored-by: Caleb Ouellette <calebo@allenai.org>