Fix message interactions on previous messages (#1128)
## Description
Closes https://github.com/allenai/playground-issues-repo/issues/544

Noticed this while I was testing some streaming. Turned out to be a
quick fix, so I went ahead and made the fix.

With some of the comparison updates our messages in a thread started
shifting around because their interactions were no longer being
rendered. This PR makes it so that MessageInteractions for old messages
stick around while a new message is streaming.

## Testing Plan

Tested locally.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/440dcd2e-0205-404e-9345-12bbe12f27bf

### After


https://github.com/user-attachments/assets/9ab7dadb-f6bf-45f1-bf9d-e0f5c6b61dd5

---------

Co-authored-by: Caleb Ouellette <calebo@allenai.org>